@@ -1,13 +1,13 @@
 {
   "name": "gerencianet/gerencianet-sdk-php",
   "description": "GN API SDK PHP",
-  "version": "1.0.7",
+  "version": "2.0.0",
   "require": {
-    "guzzlehttp/guzzle": "5.3.0",
-    "php": ">=5.4.0",
-    "codeclimate/php-test-reporter": "^0.1.2"
+    "guzzlehttp/guzzle": "^6.0.0",
+    "php": ">=5.5.0"
   },
   "require-dev": {
+    "codeclimate/php-test-reporter": "^0.1.2",
     "phpunit/phpunit": "~4.6"
   },
   "license": "MIT",
