@@ -2,6 +2,6 @@
 
 // autoload.php @generated by Composer
 
-require_once __DIR__ . '/composer' . '/autoload_real.php';
+require_once __DIR__ . '/composer/autoload_real.php';
 
-return ComposerAutoloaderInit2093506e4ada662900c429e9585b38ee::getLoader();
+return ComposerAutoloaderInitb46737289d75b56207057dd01dec455c::getLoader();
