@@ -2,7 +2,7 @@
 
 // autoload_real.php @generated by Composer
 
-class ComposerAutoloaderInit2093506e4ada662900c429e9585b38ee
+class ComposerAutoloaderInitb46737289d75b56207057dd01dec455c
 {
     private static $loader;
 
@@ -19,37 +19,48 @@ class ComposerAutoloaderInit2093506e4ada662900c429e9585b38ee
             return self::$loader;
         }
 
-        spl_autoload_register(array('ComposerAutoloaderInit2093506e4ada662900c429e9585b38ee', 'loadClassLoader'), true, true);
+        spl_autoload_register(array('ComposerAutoloaderInitb46737289d75b56207057dd01dec455c', 'loadClassLoader'), true, true);
         self::$loader = $loader = new \Composer\Autoload\ClassLoader();
-        spl_autoload_unregister(array('ComposerAutoloaderInit2093506e4ada662900c429e9585b38ee', 'loadClassLoader'));
+        spl_autoload_unregister(array('ComposerAutoloaderInitb46737289d75b56207057dd01dec455c', 'loadClassLoader'));
 
-        $map = require __DIR__ . '/autoload_namespaces.php';
-        foreach ($map as $namespace => $path) {
-            $loader->set($namespace, $path);
-        }
+        $useStaticLoader = PHP_VERSION_ID >= 50600 && !defined('HHVM_VERSION') && (!function_exists('zend_loader_file_encoded') || !zend_loader_file_encoded());
+        if ($useStaticLoader) {
+            require_once __DIR__ . '/autoload_static.php';
 
-        $map = require __DIR__ . '/autoload_psr4.php';
-        foreach ($map as $namespace => $path) {
-            $loader->setPsr4($namespace, $path);
-        }
+            call_user_func(\Composer\Autoload\ComposerStaticInitb46737289d75b56207057dd01dec455c::getInitializer($loader));
+        } else {
+            $map = require __DIR__ . '/autoload_namespaces.php';
+            foreach ($map as $namespace => $path) {
+                $loader->set($namespace, $path);
+            }
 
-        $classMap = require __DIR__ . '/autoload_classmap.php';
-        if ($classMap) {
-            $loader->addClassMap($classMap);
+            $map = require __DIR__ . '/autoload_psr4.php';
+            foreach ($map as $namespace => $path) {
+                $loader->setPsr4($namespace, $path);
+            }
+
+            $classMap = require __DIR__ . '/autoload_classmap.php';
+            if ($classMap) {
+                $loader->addClassMap($classMap);
+            }
         }
 
         $loader->register(true);
 
-        $includeFiles = require __DIR__ . '/autoload_files.php';
+        if ($useStaticLoader) {
+            $includeFiles = Composer\Autoload\ComposerStaticInitb46737289d75b56207057dd01dec455c::$files;
+        } else {
+            $includeFiles = require __DIR__ . '/autoload_files.php';
+        }
         foreach ($includeFiles as $fileIdentifier => $file) {
-            composerRequire2093506e4ada662900c429e9585b38ee($fileIdentifier, $file);
+            composerRequireb46737289d75b56207057dd01dec455c($fileIdentifier, $file);
         }
 
         return $loader;
     }
 }
 
-function composerRequire2093506e4ada662900c429e9585b38ee($fileIdentifier, $file)
+function composerRequireb46737289d75b56207057dd01dec455c($fileIdentifier, $file)
 {
     if (empty($GLOBALS['__composer_autoload_files'][$fileIdentifier])) {
         require $file;
