@@ -1,322 +1,242 @@
 <?php
 namespace GuzzleHttp;
 
-use GuzzleHttp\Event\HasEmitterTrait;
-use GuzzleHttp\Message\MessageFactory;
-use GuzzleHttp\Message\MessageFactoryInterface;
-use GuzzleHttp\Message\RequestInterface;
-use GuzzleHttp\Message\FutureResponse;
-use GuzzleHttp\Ring\Core;
-use GuzzleHttp\Ring\Future\FutureInterface;
-use GuzzleHttp\Exception\RequestException;
-use React\Promise\FulfilledPromise;
-use React\Promise\RejectedPromise;
+use GuzzleHttp\Cookie\CookieJar;
+use GuzzleHttp\Promise;
+use GuzzleHttp\Psr7;
+use Psr\Http\Message\UriInterface;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
 
 /**
- * HTTP client
+ * @method ResponseInterface get(string|UriInterface $uri, array $options = [])
+ * @method ResponseInterface head(string|UriInterface $uri, array $options = [])
+ * @method ResponseInterface put(string|UriInterface $uri, array $options = [])
+ * @method ResponseInterface post(string|UriInterface $uri, array $options = [])
+ * @method ResponseInterface patch(string|UriInterface $uri, array $options = [])
+ * @method ResponseInterface delete(string|UriInterface $uri, array $options = [])
+ * @method Promise\PromiseInterface getAsync(string|UriInterface $uri, array $options = [])
+ * @method Promise\PromiseInterface headAsync(string|UriInterface $uri, array $options = [])
+ * @method Promise\PromiseInterface putAsync(string|UriInterface $uri, array $options = [])
+ * @method Promise\PromiseInterface postAsync(string|UriInterface $uri, array $options = [])
+ * @method Promise\PromiseInterface patchAsync(string|UriInterface $uri, array $options = [])
+ * @method Promise\PromiseInterface deleteAsync(string|UriInterface $uri, array $options = [])
  */
 class Client implements ClientInterface
 {
-    use HasEmitterTrait;
-
-    /** @var MessageFactoryInterface Request factory used by the client */
-    private $messageFactory;
-
-    /** @var Url Base URL of the client */
-    private $baseUrl;
-
     /** @var array Default request options */
-    private $defaults;
-
-    /** @var callable Request state machine */
-    private $fsm;
+    private $config;
 
     /**
      * Clients accept an array of constructor parameters.
      *
-     * Here's an example of creating a client using an URI template for the
-     * client's base_url and an array of default request options to apply
-     * to each request:
+     * Here's an example of creating a client using a base_uri and an array of
+     * default request options to apply to each request:
      *
      *     $client = new Client([
-     *         'base_url' => [
-     *              'http://www.foo.com/{version}/',
-     *              ['version' => '123']
-     *          ],
-     *         'defaults' => [
-     *             'timeout'         => 10,
-     *             'allow_redirects' => false,
-     *             'proxy'           => '192.168.16.1:10'
-     *         ]
+     *         'base_uri'        => 'http://www.foo.com/1.0/',
+     *         'timeout'         => 0,
+     *         'allow_redirects' => false,
+     *         'proxy'           => '192.168.16.1:10'
      *     ]);
      *
-     * @param array $config Client configuration settings
-     *     - base_url: Base URL of the client that is merged into relative URLs.
-     *       Can be a string or an array that contains a URI template followed
-     *       by an associative array of expansion variables to inject into the
-     *       URI template.
-     *     - handler: callable RingPHP handler used to transfer requests
-     *     - message_factory: Factory used to create request and response object
-     *     - defaults: Default request options to apply to each request
-     *     - emitter: Event emitter used for request events
-     *     - fsm: (internal use only) The request finite state machine. A
-     *       function that accepts a transaction and optional final state. The
-     *       function is responsible for transitioning a request through its
-     *       lifecycle events.
+     * Client configuration settings include the following options:
+     *
+     * - handler: (callable) Function that transfers HTTP requests over the
+     *   wire. The function is called with a Psr7\Http\Message\RequestInterface
+     *   and array of transfer options, and must return a
+     *   GuzzleHttp\Promise\PromiseInterface that is fulfilled with a
+     *   Psr7\Http\Message\ResponseInterface on success. "handler" is a
+     *   constructor only option that cannot be overridden in per/request
+     *   options. If no handler is provided, a default handler will be created
+     *   that enables all of the request options below by attaching all of the
+     *   default middleware to the handler.
+     * - base_uri: (string|UriInterface) Base URI of the client that is merged
+     *   into relative URIs. Can be a string or instance of UriInterface.
+     * - **: any request option
+     *
+     * @param array $config Client configuration settings.
+     *
+     * @see \GuzzleHttp\RequestOptions for a list of available request options.
      */
     public function __construct(array $config = [])
     {
-        $this->configureBaseUrl($config);
+        if (!isset($config['handler'])) {
+            $config['handler'] = HandlerStack::create();
+        } elseif (!is_callable($config['handler'])) {
+            throw new \InvalidArgumentException('handler must be a callable');
+        }
+
+        // Convert the base_uri to a UriInterface
+        if (isset($config['base_uri'])) {
+            $config['base_uri'] = Psr7\uri_for($config['base_uri']);
+        }
+
         $this->configureDefaults($config);
+    }
 
-        if (isset($config['emitter'])) {
-            $this->emitter = $config['emitter'];
+    public function __call($method, $args)
+    {
+        if (count($args) < 1) {
+            throw new \InvalidArgumentException('Magic request methods require a URI and optional options array');
         }
 
-        $this->messageFactory = isset($config['message_factory'])
-            ? $config['message_factory']
-            : new MessageFactory();
+        $uri = $args[0];
+        $opts = isset($args[1]) ? $args[1] : [];
 
-        if (isset($config['fsm'])) {
-            $this->fsm = $config['fsm'];
-        } else {
-            if (isset($config['handler'])) {
-                $handler = $config['handler'];
-            } elseif (isset($config['adapter'])) {
-                $handler = $config['adapter'];
-            } else {
-                $handler = Utils::getDefaultHandler();
-            }
-            $this->fsm = new RequestFsm($handler, $this->messageFactory);
-        }
+        return substr($method, -5) === 'Async'
+            ? $this->requestAsync(substr($method, 0, -5), $uri, $opts)
+            : $this->request($method, $uri, $opts);
     }
 
-    public function getDefaultOption($keyOrPath = null)
+    public function sendAsync(RequestInterface $request, array $options = [])
     {
-        return $keyOrPath === null
-            ? $this->defaults
-            : Utils::getPath($this->defaults, $keyOrPath);
-    }
+        // Merge the base URI into the request URI if needed.
+        $options = $this->prepareDefaults($options);
 
-    public function setDefaultOption($keyOrPath, $value)
-    {
-        Utils::setPath($this->defaults, $keyOrPath, $value);
-    }
-
-    public function getBaseUrl()
-    {
-        return (string) $this->baseUrl;
-    }
-
-    public function createRequest($method, $url = null, array $options = [])
-    {
-        $options = $this->mergeDefaults($options);
-        // Use a clone of the client's emitter
-        $options['config']['emitter'] = clone $this->getEmitter();
-        $url = $url || (is_string($url) && strlen($url))
-            ? $this->buildUrl($url)
-            : (string) $this->baseUrl;
-
-        return $this->messageFactory->createRequest($method, $url, $options);
-    }
-
-    public function get($url = null, $options = [])
-    {
-        return $this->send($this->createRequest('GET', $url, $options));
-    }
-
-    public function head($url = null, array $options = [])
-    {
-        return $this->send($this->createRequest('HEAD', $url, $options));
-    }
-
-    public function delete($url = null, array $options = [])
-    {
-        return $this->send($this->createRequest('DELETE', $url, $options));
-    }
-
-    public function put($url = null, array $options = [])
-    {
-        return $this->send($this->createRequest('PUT', $url, $options));
-    }
-
-    public function patch($url = null, array $options = [])
-    {
-        return $this->send($this->createRequest('PATCH', $url, $options));
-    }
-
-    public function post($url = null, array $options = [])
-    {
-        return $this->send($this->createRequest('POST', $url, $options));
-    }
-
-    public function options($url = null, array $options = [])
-    {
-        return $this->send($this->createRequest('OPTIONS', $url, $options));
-    }
-
-    public function send(RequestInterface $request)
-    {
-        $isFuture = $request->getConfig()->get('future');
-        $trans = new Transaction($this, $request, $isFuture);
-        $fn = $this->fsm;
-
-        try {
-            $fn($trans);
-            if ($isFuture) {
-                // Turn the normal response into a future if needed.
-                return $trans->response instanceof FutureInterface
-                    ? $trans->response
-                    : new FutureResponse(new FulfilledPromise($trans->response));
-            }
-            // Resolve deep futures if this is not a future
-            // transaction. This accounts for things like retries
-            // that do not have an immediate side-effect.
-            while ($trans->response instanceof FutureInterface) {
-                $trans->response = $trans->response->wait();
-            }
-            return $trans->response;
-        } catch (\Exception $e) {
-            if ($isFuture) {
-                // Wrap the exception in a promise
-                return new FutureResponse(new RejectedPromise($e));
-            }
-            throw RequestException::wrapException($trans->request, $e);
-        }
-    }
-
-    /**
-     * Get an array of default options to apply to the client
-     *
-     * @return array
-     */
-    protected function getDefaultOptions()
-    {
-        $settings = [
-            'allow_redirects' => true,
-            'exceptions'      => true,
-            'decode_content'  => true,
-            'verify'          => true
-        ];
-
-        // Use the standard Linux HTTP_PROXY and HTTPS_PROXY if set
-        if ($proxy = getenv('HTTP_PROXY')) {
-            $settings['proxy']['http'] = $proxy;
-        }
-
-        if ($proxy = getenv('HTTPS_PROXY')) {
-            $settings['proxy']['https'] = $proxy;
-        }
-
-        return $settings;
-    }
-
-    /**
-     * Expand a URI template and inherit from the base URL if it's relative
-     *
-     * @param string|array $url URL or an array of the URI template to expand
-     *                          followed by a hash of template varnames.
-     * @return string
-     * @throws \InvalidArgumentException
-     */
-    private function buildUrl($url)
-    {
-        // URI template (absolute or relative)
-        if (!is_array($url)) {
-            return strpos($url, '://')
-                ? (string) $url
-                : (string) $this->baseUrl->combine($url);
-        }
-
-        if (!isset($url[1])) {
-            throw new \InvalidArgumentException('You must provide a hash of '
-                . 'varname options in the second element of a URL array.');
-        }
-
-        // Absolute URL
-        if (strpos($url[0], '://')) {
-            return Utils::uriTemplate($url[0], $url[1]);
-        }
-
-        // Combine the relative URL with the base URL
-        return (string) $this->baseUrl->combine(
-            Utils::uriTemplate($url[0], $url[1])
+        return $this->transfer(
+            $request->withUri($this->buildUri($request->getUri(), $options), $request->hasHeader('Host')),
+            $options
         );
     }
 
-    private function configureBaseUrl(&$config)
+    public function send(RequestInterface $request, array $options = [])
     {
-        if (!isset($config['base_url'])) {
-            $this->baseUrl = new Url('', '');
-        } elseif (!is_array($config['base_url'])) {
-            $this->baseUrl = Url::fromString($config['base_url']);
-        } elseif (count($config['base_url']) < 2) {
-            throw new \InvalidArgumentException('You must provide a hash of '
-                . 'varname options in the second element of a base_url array.');
-        } else {
-            $this->baseUrl = Url::fromString(
-                Utils::uriTemplate(
-                    $config['base_url'][0],
-                    $config['base_url'][1]
-                )
-            );
-            $config['base_url'] = (string) $this->baseUrl;
-        }
+        $options[RequestOptions::SYNCHRONOUS] = true;
+        return $this->sendAsync($request, $options)->wait();
     }
 
-    private function configureDefaults($config)
+    public function requestAsync($method, $uri = '', array $options = [])
     {
-        if (!isset($config['defaults'])) {
-            $this->defaults = $this->getDefaultOptions();
-        } else {
-            $this->defaults = array_replace(
-                $this->getDefaultOptions(),
-                $config['defaults']
-            );
+        $options = $this->prepareDefaults($options);
+        // Remove request modifying parameter because it can be done up-front.
+        $headers = isset($options['headers']) ? $options['headers'] : [];
+        $body = isset($options['body']) ? $options['body'] : null;
+        $version = isset($options['version']) ? $options['version'] : '1.1';
+        // Merge the URI into the base URI.
+        $uri = $this->buildUri($uri, $options);
+        if (is_array($body)) {
+            $this->invalidBody();
+        }
+        $request = new Psr7\Request($method, $uri, $headers, $body, $version);
+        // Remove the option so that they are not doubly-applied.
+        unset($options['headers'], $options['body'], $options['version']);
+
+        return $this->transfer($request, $options);
+    }
+
+    public function request($method, $uri = '', array $options = [])
+    {
+        $options[RequestOptions::SYNCHRONOUS] = true;
+        return $this->requestAsync($method, $uri, $options)->wait();
+    }
+
+    public function getConfig($option = null)
+    {
+        return $option === null
+            ? $this->config
+            : (isset($this->config[$option]) ? $this->config[$option] : null);
+    }
+
+    private function buildUri($uri, array $config)
+    {
+        // for BC we accept null which would otherwise fail in uri_for
+        $uri = Psr7\uri_for($uri === null ? '' : $uri);
+
+        if (isset($config['base_uri'])) {
+            $uri = Psr7\UriResolver::resolve(Psr7\uri_for($config['base_uri']), $uri);
         }
 
-        // Add the default user-agent header
-        if (!isset($this->defaults['headers'])) {
-            $this->defaults['headers'] = [
-                'User-Agent' => Utils::getDefaultUserAgent()
-            ];
-        } elseif (!Core::hasHeader($this->defaults, 'User-Agent')) {
-            // Add the User-Agent header if one was not already set
-            $this->defaults['headers']['User-Agent'] = Utils::getDefaultUserAgent();
+        return $uri->getScheme() === '' && $uri->getHost() !== '' ? $uri->withScheme('http') : $uri;
+    }
+
+    /**
+     * Configures the default options for a client.
+     *
+     * @param array $config
+     */
+    private function configureDefaults(array $config)
+    {
+        $defaults = [
+            'allow_redirects' => RedirectMiddleware::$defaultSettings,
+            'http_errors'     => true,
+            'decode_content'  => true,
+            'verify'          => true,
+            'cookies'         => false
+        ];
+
+        // Use the standard Linux HTTP_PROXY and HTTPS_PROXY if set.
+
+        // We can only trust the HTTP_PROXY environment variable in a CLI
+        // process due to the fact that PHP has no reliable mechanism to
+        // get environment variables that start with "HTTP_".
+        if (php_sapi_name() == 'cli' && getenv('HTTP_PROXY')) {
+            $defaults['proxy']['http'] = getenv('HTTP_PROXY');
+        }
+
+        if ($proxy = getenv('HTTPS_PROXY')) {
+            $defaults['proxy']['https'] = $proxy;
+        }
+
+        if ($noProxy = getenv('NO_PROXY')) {
+            $cleanedNoProxy = str_replace(' ', '', $noProxy);
+            $defaults['proxy']['no'] = explode(',', $cleanedNoProxy);
+        }
+
+        $this->config = $config + $defaults;
+
+        if (!empty($config['cookies']) && $config['cookies'] === true) {
+            $this->config['cookies'] = new CookieJar();
+        }
+
+        // Add the default user-agent header.
+        if (!isset($this->config['headers'])) {
+            $this->config['headers'] = ['User-Agent' => default_user_agent()];
+        } else {
+            // Add the User-Agent header if one was not already set.
+            foreach (array_keys($this->config['headers']) as $name) {
+                if (strtolower($name) === 'user-agent') {
+                    return;
+                }
+            }
+            $this->config['headers']['User-Agent'] = default_user_agent();
         }
     }
 
     /**
-     * Merges default options into the array passed by reference.
+     * Merges default options into the array.
      *
      * @param array $options Options to modify by reference
      *
      * @return array
      */
-    private function mergeDefaults($options)
+    private function prepareDefaults($options)
     {
-        $defaults = $this->defaults;
+        $defaults = $this->config;
 
-        // Case-insensitively merge in default headers if both defaults and
-        // options have headers specified.
-        if (!empty($defaults['headers']) && !empty($options['headers'])) {
-            // Create a set of lowercased keys that are present.
-            $lkeys = [];
-            foreach (array_keys($options['headers']) as $k) {
-                $lkeys[strtolower($k)] = true;
-            }
-            // Merge in lowercase default keys when not present in above set.
-            foreach ($defaults['headers'] as $key => $value) {
-                if (!isset($lkeys[strtolower($key)])) {
-                    $options['headers'][$key] = $value;
-                }
-            }
-            // No longer need to merge in headers.
+        if (!empty($defaults['headers'])) {
+            // Default headers are only added if they are not present.
+            $defaults['_conditional'] = $defaults['headers'];
             unset($defaults['headers']);
         }
 
-        $result = array_replace_recursive($defaults, $options);
-        foreach ($options as $k => $v) {
+        // Special handling for headers is required as they are added as
+        // conditional headers and as headers passed to a request ctor.
+        if (array_key_exists('headers', $options)) {
+            // Allows default headers to be unset.
+            if ($options['headers'] === null) {
+                $defaults['_conditional'] = null;
+                unset($options['headers']);
+            } elseif (!is_array($options['headers'])) {
+                throw new \InvalidArgumentException('headers must be an array');
+            }
+        }
+
+        // Shallow merge defaults underneath options.
+        $result = $options + $defaults;
+
+        // Remove null values.
+        foreach ($result as $k => $v) {
             if ($v === null) {
                 unset($result[$k]);
             }
@@ -326,27 +246,169 @@ class Client implements ClientInterface
     }
 
     /**
-     * @deprecated Use {@see GuzzleHttp\Pool} instead.
-     * @see GuzzleHttp\Pool
+     * Transfers the given request and applies request options.
+     *
+     * The URI of the request is not modified and the request options are used
+     * as-is without merging in default options.
+     *
+     * @param RequestInterface $request
+     * @param array            $options
+     *
+     * @return Promise\PromiseInterface
      */
-    public function sendAll($requests, array $options = [])
+    private function transfer(RequestInterface $request, array $options)
     {
-        Pool::send($this, $requests, $options);
+        // save_to -> sink
+        if (isset($options['save_to'])) {
+            $options['sink'] = $options['save_to'];
+            unset($options['save_to']);
+        }
+
+        // exceptions -> http_errors
+        if (isset($options['exceptions'])) {
+            $options['http_errors'] = $options['exceptions'];
+            unset($options['exceptions']);
+        }
+
+        $request = $this->applyOptions($request, $options);
+        $handler = $options['handler'];
+
+        try {
+            return Promise\promise_for($handler($request, $options));
+        } catch (\Exception $e) {
+            return Promise\rejection_for($e);
+        }
     }
 
     /**
-     * @deprecated Use GuzzleHttp\Utils::getDefaultHandler
+     * Applies the array of request options to a request.
+     *
+     * @param RequestInterface $request
+     * @param array            $options
+     *
+     * @return RequestInterface
      */
-    public static function getDefaultHandler()
+    private function applyOptions(RequestInterface $request, array &$options)
     {
-        return Utils::getDefaultHandler();
+        $modify = [];
+
+        if (isset($options['form_params'])) {
+            if (isset($options['multipart'])) {
+                throw new \InvalidArgumentException('You cannot use '
+                    . 'form_params and multipart at the same time. Use the '
+                    . 'form_params option if you want to send application/'
+                    . 'x-www-form-urlencoded requests, and the multipart '
+                    . 'option to send multipart/form-data requests.');
+            }
+            $options['body'] = http_build_query($options['form_params'], '', '&');
+            unset($options['form_params']);
+            $options['_conditional']['Content-Type'] = 'application/x-www-form-urlencoded';
+        }
+
+        if (isset($options['multipart'])) {
+            $options['body'] = new Psr7\MultipartStream($options['multipart']);
+            unset($options['multipart']);
+        }
+
+        if (isset($options['json'])) {
+            $options['body'] = \GuzzleHttp\json_encode($options['json']);
+            unset($options['json']);
+            $options['_conditional']['Content-Type'] = 'application/json';
+        }
+
+        if (!empty($options['decode_content'])
+            && $options['decode_content'] !== true
+        ) {
+            $modify['set_headers']['Accept-Encoding'] = $options['decode_content'];
+        }
+
+        if (isset($options['headers'])) {
+            if (isset($modify['set_headers'])) {
+                $modify['set_headers'] = $options['headers'] + $modify['set_headers'];
+            } else {
+                $modify['set_headers'] = $options['headers'];
+            }
+            unset($options['headers']);
+        }
+
+        if (isset($options['body'])) {
+            if (is_array($options['body'])) {
+                $this->invalidBody();
+            }
+            $modify['body'] = Psr7\stream_for($options['body']);
+            unset($options['body']);
+        }
+
+        if (!empty($options['auth']) && is_array($options['auth'])) {
+            $value = $options['auth'];
+            $type = isset($value[2]) ? strtolower($value[2]) : 'basic';
+            switch ($type) {
+                case 'basic':
+                    $modify['set_headers']['Authorization'] = 'Basic '
+                        . base64_encode("$value[0]:$value[1]");
+                    break;
+                case 'digest':
+                    // @todo: Do not rely on curl
+                    $options['curl'][CURLOPT_HTTPAUTH] = CURLAUTH_DIGEST;
+                    $options['curl'][CURLOPT_USERPWD] = "$value[0]:$value[1]";
+                    break;
+                case 'ntlm':
+                    $options['curl'][CURLOPT_HTTPAUTH] = CURLAUTH_NTLM;
+                    $options['curl'][CURLOPT_USERPWD] = "$value[0]:$value[1]";
+                    break;
+            }
+        }
+
+        if (isset($options['query'])) {
+            $value = $options['query'];
+            if (is_array($value)) {
+                $value = http_build_query($value, null, '&', PHP_QUERY_RFC3986);
+            }
+            if (!is_string($value)) {
+                throw new \InvalidArgumentException('query must be a string or array');
+            }
+            $modify['query'] = $value;
+            unset($options['query']);
+        }
+
+        // Ensure that sink is not an invalid value.
+        if (isset($options['sink'])) {
+            // TODO: Add more sink validation?
+            if (is_bool($options['sink'])) {
+                throw new \InvalidArgumentException('sink must not be a boolean');
+            }
+        }
+
+        $request = Psr7\modify_request($request, $modify);
+        if ($request->getBody() instanceof Psr7\MultipartStream) {
+            // Use a multipart/form-data POST if a Content-Type is not set.
+            $options['_conditional']['Content-Type'] = 'multipart/form-data; boundary='
+                . $request->getBody()->getBoundary();
+        }
+
+        // Merge in conditional headers if they are not present.
+        if (isset($options['_conditional'])) {
+            // Build up the changes so it's in a single clone of the message.
+            $modify = [];
+            foreach ($options['_conditional'] as $k => $v) {
+                if (!$request->hasHeader($k)) {
+                    $modify['set_headers'][$k] = $v;
+                }
+            }
+            $request = Psr7\modify_request($request, $modify);
+            // Don't pass this internal value along to middleware/handlers.
+            unset($options['_conditional']);
+        }
+
+        return $request;
     }
 
-    /**
-     * @deprecated Use GuzzleHttp\Utils::getDefaultUserAgent
-     */
-    public static function getDefaultUserAgent()
+    private function invalidBody()
     {
-        return Utils::getDefaultUserAgent();
+        throw new \InvalidArgumentException('Passing in the "body" request '
+            . 'option as an array to send a POST request has been deprecated. '
+            . 'Please use the "form_params" request option to send a '
+            . 'application/x-www-form-urlencoded request, or the "multipart" '
+            . 'request option to send a multipart/form-data request.');
     }
 }
