@@ -1,8 +1,6 @@
 <?php
 namespace GuzzleHttp\Cookie;
 
-use GuzzleHttp\Utils;
-
 /**
  * Persists non-session cookies using a JSON formatted file
  */
@@ -11,16 +9,22 @@ class FileCookieJar extends CookieJar
     /** @var string filename */
     private $filename;
 
+    /** @var bool Control whether to persist session cookies or not. */
+    private $storeSessionCookies;
+
     /**
      * Create a new FileCookieJar object
      *
-     * @param string $cookieFile File to store the cookie data
+     * @param string $cookieFile        File to store the cookie data
+     * @param bool $storeSessionCookies Set to true to store session cookies
+     *                                  in the cookie jar.
      *
      * @throws \RuntimeException if the file cannot be found or created
      */
-    public function __construct($cookieFile)
+    public function __construct($cookieFile, $storeSessionCookies = false)
     {
         $this->filename = $cookieFile;
+        $this->storeSessionCookies = $storeSessionCookies;
 
         if (file_exists($cookieFile)) {
             $this->load($cookieFile);
@@ -45,15 +49,15 @@ class FileCookieJar extends CookieJar
     {
         $json = [];
         foreach ($this as $cookie) {
-            if ($cookie->getExpires() && !$cookie->getDiscard()) {
+            /** @var SetCookie $cookie */
+            if (CookieJar::shouldPersist($cookie, $this->storeSessionCookies)) {
                 $json[] = $cookie->toArray();
             }
         }
 
-        if (false === file_put_contents($filename, json_encode($json))) {
-            // @codeCoverageIgnoreStart
+        $jsonStr = \GuzzleHttp\json_encode($json);
+        if (false === file_put_contents($filename, $jsonStr)) {
             throw new \RuntimeException("Unable to save file {$filename}");
-            // @codeCoverageIgnoreEnd
         }
     }
 
@@ -69,14 +73,14 @@ class FileCookieJar extends CookieJar
     {
         $json = file_get_contents($filename);
         if (false === $json) {
-            // @codeCoverageIgnoreStart
             throw new \RuntimeException("Unable to load file {$filename}");
-            // @codeCoverageIgnoreEnd
+        } elseif ($json === '') {
+            return;
         }
 
-        $data = Utils::jsonDecode($json, true);
+        $data = \GuzzleHttp\json_decode($json, true);
         if (is_array($data)) {
-            foreach (Utils::jsonDecode($json, true) as $cookie) {
+            foreach (json_decode($json, true) as $cookie) {
                 $this->setCookie(new SetCookie($cookie));
             }
         } elseif (strlen($data)) {
