@@ -1,8 +1,6 @@
 <?php
 namespace GuzzleHttp\Cookie;
 
-use GuzzleHttp\Utils;
-
 /**
  * Persists cookies in the client session
  */
@@ -10,15 +8,22 @@ class SessionCookieJar extends CookieJar
 {
     /** @var string session key */
     private $sessionKey;
+    
+    /** @var bool Control whether to persist session cookies or not. */
+    private $storeSessionCookies;
 
     /**
      * Create a new SessionCookieJar object
      *
-     * @param string $sessionKey Session key name to store the cookie data in session
+     * @param string $sessionKey        Session key name to store the cookie 
+     *                                  data in session
+     * @param bool $storeSessionCookies Set to true to store session cookies
+     *                                  in the cookie jar.
      */
-    public function __construct($sessionKey)
+    public function __construct($sessionKey, $storeSessionCookies = false)
     {
         $this->sessionKey = $sessionKey;
+        $this->storeSessionCookies = $storeSessionCookies;
         $this->load();
     }
 
@@ -37,7 +42,8 @@ class SessionCookieJar extends CookieJar
     {
         $json = [];
         foreach ($this as $cookie) {
-            if ($cookie->getExpires() && !$cookie->getDiscard()) {
+            /** @var SetCookie $cookie */
+            if (CookieJar::shouldPersist($cookie, $this->storeSessionCookies)) {
                 $json[] = $cookie->toArray();
             }
         }
@@ -50,11 +56,10 @@ class SessionCookieJar extends CookieJar
      */
     protected function load()
     {
-        $cookieJar = isset($_SESSION[$this->sessionKey])
-            ? $_SESSION[$this->sessionKey]
-            : null;
-
-        $data = Utils::jsonDecode($cookieJar, true);
+        if (!isset($_SESSION[$this->sessionKey])) {
+            return;
+        }
+        $data = json_decode($_SESSION[$this->sessionKey], true);
         if (is_array($data)) {
             foreach ($data as $cookie) {
                 $this->setCookie(new SetCookie($cookie));
