@@ -1,4 +1,4 @@
 <?php
 namespace GuzzleHttp\Exception;
 
-class TransferException extends \RuntimeException {}
+class TransferException extends \RuntimeException implements GuzzleException {}
